Create Bootstrap_RemoteDesktopServers_on_AWS_ASG.yml
This document can be used along with AWS ASG to bring up Fully function Remote Desktop hosts for an RDS farm.  It joins them to the domain, names them, and then after a reboot, it adds them to the RDS deployment and sets them as part of the published Application or Desktop.  See code for more description and comments.